검 이펙트 구현 & 어드레서블 에셋 추가

+++ Assets/01.Scripts/Utils/Define.cs
@@ -52,5 +52,5 @@ public class Define
     public const int ELEMENT_MARBLE_COUNT = 3;
     #endregion
 
-    public const float FIXED_FPS = 50f;
+    public const int FIXED_FPS = 50;
 }
